# Conflicts:
#	Source/CMakeLists.txt


Former-commit-id: 2e7e9d7e1da964c9fbd2b6a340e3ae5701349b71

+++ Source/MBansheeEditor/Inspectors/AudioSourceInspector.cs
@@ -92,7 +92,7 @@ namespace BansheeEditor
                 ConfirmModify();
             };
 
-            priorityField.OnChanged += x => { source.Priority = x; MarkAsModified(); };
+            priorityField.OnChanged += x => { source.Priority = (uint)x; MarkAsModified(); };
             priorityField.OnConfirmed += ConfirmModify;
             priorityField.OnFocusLost += ConfirmModify;
